requirements.txt
selenium
python-decouple
chromedriver-py
numpy